Ensure passed canvas abides by render options
Engine.create takes a canvas element or creates a new canvas using the
given options. While the properties width and height in options do apply
to the created canvas, they do not apply to the passed canvas. These set
options are under render.canvas, render.options.width and
render.options.height. This issue was fixed by setting the canvas width
and height only when a canvas element is directly provided.