Reverse vertices order from counterclockwise to clockwise if one scale is negative.
Regarding: https://github.com/liabru/matter-js/issues/443
I was having an issue with bodies that were flipped over X wobbling on edges. But not in their original scale. I saw the above issue and thought that might be the issue. So I created this solution which solved the issue of the body wobbling on the edge and some minor other issues I was noticing.