Added lockRotation for sprite rendering
This simple yet intuitive addition enables you to toggle whether or not to rotate the sprite along with the rigidbody.

I was working on a simulation where I have different sprite sheets for different angular positions for a body but due to this not being an option I had to hack my way my directly drawing in the canvas on 'afterRender' event on top of the existing rigidbody drawing. This will atleast take care of offseting that external operation.

In the future I also plan to add a function for rendering a series of sprites on given intervals based on a map of angular position--something which I have been doing through events.